// FundooNotesApp/src/app/models/register.model.ts
export class RegisterModel {
    firstName: String;
    lastName: String;
    email: String;
    password: String;
    phoneNumber: Number;
    service: string;
}

// "firstName": "string",
// "lastName": "string",
// "phoneNumber": "string",
// "imageUrl": "string",
// "role": "string",
// "service": "string",
// "createdDate": "2019-02-09T19:35:18.058Z",
// "modifiedDate": "2019-02-09T19:35:18.058Z",
// "addresses": [
//   {}
// ],
// "realm": "string",
// "username": "string",
// "email": "string",
// "emailVerified": true,
// "id": "string"
// }